# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Listou API
  version: 0.1.0
  description: >
    REST API do Listou. Autenticação por cookie de sessão `listou_session` (HttpOnly).
    Todos os erros seguem o envelope `{ "error": { "code", "message", "requestId", "fields?" } }`.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/register:
    post:
      summary: Cria conta e inicia sessão
      tags: [auth]
      requestBody:
        {
          required: true,
          content: { application/json: { schema: { $ref: "#/components/schemas/Register" } } },
        }
      responses:
        "201":
          {
            description: Conta criada (Set-Cookie listou_session),
            content:
              { application/json: { schema: { $ref: "#/components/schemas/UserEnvelope" } } },
          }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/auth/login:
    post:
      summary: Login com e-mail e senha
      tags: [auth]
      requestBody:
        {
          required: true,
          content: { application/json: { schema: { $ref: "#/components/schemas/Login" } } },
        }
      responses:
        "200":
          {
            description: ok,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/UserEnvelope" } } },
          }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/v1/me:
    {
      get:
        {
          summary: Usuário atual,
          tags: [auth],
          security: [{ session: [] }],
          responses:
            {
              "200":
                {
                  description: ok,
                  content:
                    { application/json: { schema: { $ref: "#/components/schemas/UserEnvelope" } } },
                },
              "401": { $ref: "#/components/responses/Error" },
            },
        },
    }
...
  /api/v1/products/search:
    get:
      summary: Busca unificada nos providers habilitados
      tags: [catalog]
      security: [{ session: [] }]
      parameters:
        - {
            name: q,
            in: query,
            required: true,
            schema: { type: string, minLength: 2, maxLength: 100 },
          }
        - {
            name: sort,
            in: query,
            schema: { type: string, enum: [relevance, price_asc, price_desc] },
          }
        - { name: limit, in: query, schema: { type: integer, maximum: 40 } }
      responses: { "200": { description: ok } }
  /api/v1/products/import:
    {
      post:
        {
          summary: Persiste produto + ofertas de um provider,
          tags: [catalog],
          security: [{ session: [] }],
          responses: { "200": { description: ok } },
        },
    }
  /api/v1/products/link-preview:
    {
      post:
        {
          summary: Lê título e imagem públicos de uma URL de produto (sem preço),
          tags: [catalog],
          security: [{ session: [] }],
          responses:
            { "200": { description: ok }, "400": { description: INVALID_LINK } },
        },
    }
  /api/v1/products/import-link:
    {
      post:
        {
          summary: Persiste um produto a partir de uma URL colada,
          tags: [catalog],
          security: [{ session: [] }],
          responses:
            {
              "200": { description: ok },
              "400": { description: INVALID_LINK ou LINK_UNREADABLE },
            },
        },
    }
  /api/v1/admin/curated-products:
    {
      get:
        {
          summary: Lista o catálogo curado e as lojas disponíveis (somente ADMIN),
          tags: [admin],
          security: [{ session: [] }],
          responses: { "200": { description: ok }, "403": { description: FORBIDDEN } },
        },
      post:
        {
          summary: Cria um produto curado com links de afiliado (somente ADMIN),
          tags: [admin],
          security: [{ session: [] }],
          responses:
            {
              "201": { description: criado },
              "403": { description: FORBIDDEN },
              "422": { description: VALIDATION_FAILED },
            },
        },
    }
  /api/v1/admin/curated-products/{id}:
    {
      put:
        {
          summary: Substitui campos e ofertas de um produto curado (somente ADMIN),
          tags: [admin],
          parameters: [{ $ref: "#/components/parameters/id" }],
          security: [{ session: [] }],
          responses:
            {
              "200": { description: ok },
              "403": { description: FORBIDDEN },
              "404": { description: CURATED_PRODUCT_NOT_FOUND },
            },
        },
      delete:
        {
          summary: Remove um produto curado (somente ADMIN),
          tags: [admin],
          parameters: [{ $ref: "#/components/parameters/id" }],
          security: [{ session: [] }],
          responses: { "204": { description: removido }, "403": { description: FORBIDDEN } },
        },
    }
  /api/v1/products/{id}:
    {
      get:
        {
          summary: Produto e ofertas,
          tags: [catalog],
          parameters: [{ $ref: "#/components/parameters/id" }],
          security: [{ session: [] }],
          responses: { "200": { description: ok } },
        },
    }
  /api/v1/products/{id}/offers:
    {
      get:
        {
          summary: Ofertas (sem URL de loja; use goUrl),
          tags: [catalog],
          parameters: [{ $ref: "#/components/parameters/id" }],
          security: [{ session: [] }],
          responses: { "200": { description: ok } },
        },
    }
...
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: listou_session }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    Error:
      description: Erro padronizado
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, example: LIST_NOT_FOUND }
            message: { type: string }
            requestId: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
    Register:
      {
        type: object,
        required: [name, email, password],
        properties:
          {
            name: { type: string },
            email: { type: string, format: email },
            password: { type: string, minLength: 8, maxLength: 128 },
          },
      }
    Login:
      {
        type: object,
        required: [email, password],
        properties: { email: { type: string }, password: { type: string } },
      }
    UserEnvelope:
      {
        type: object,
        properties:
          {
            user:
              {
                type: object,
                properties:
                  {
                    id: { type: string },
                    email: { type: string },
                    name: { type: string },
                    role: { type: string, enum: [USER, ADMIN] },
                  },
              },
          },
      }
...